Справочник настроек

+++ config/services.php
@@ -18,7 +18,7 @@ return [
     ],
 
     'telegram' => [
-        'admin_id' => env('TELEGRAM_ADMIN_ID', 0),
+        'admin_id' => (int) env('TELEGRAM_ADMIN_ID', 0),
         'bot_token' => env('TELEGRAM_BOT_TOKEN', 'YOUR-BOT-TOKEN'),
         'webhook_secret_token' => env('TELEGRAM_WEBHOOK_SECRET_TOKEN'),
     ],
